requirements.txt
# YOLOv8 Training and Inference Requirements
# For local training and Raspberry Pi deployment

# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=9.5.0

# Data handling
pyyaml>=6.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress tracking
tqdm>=4.65.0

# MQTT and Dashboard
paho-mqtt>=1.6.0
flask>=2.3.0
flask-socketio>=5.3.0

# For Raspberry Pi (install on Pi only)
# tensorflow-lite-runtime>=2.13.0  # Uncomment on Raspberry Pi
